putting sortDates function into common date file
