+++ src/common/date.js
@@ -12,3 +12,15 @@ export function getTimeStamp(dateString) {
   const parsedDate = Date.parse(dateString);
   return parsedDate;
 }
+
+export function sortDates(aRow, bRow, key) {
+  const a = aRow[key];
+  const b = bRow[key];
+
+  if (!a || !b) {
+    return -1;
+  }
+  const dateA = a.split('-').join('');
+  const dateB = b.split('-').join('');
+  return dateA < dateB ? -1 : dateA > dateB ? 1 : 0;
+}
