+++ src/components/Absentees/Absentees.vue
@@ -37,7 +37,7 @@
       striped hover
       sort-by="date"
       sort-desc
-      :sort-compare="sortDates"
+      :sort-compare="sort"
       :items="absentees"
       :fields="absenteeFields"
       :per-page="perPage"
@@ -54,7 +54,7 @@
 import Vue from 'vue';
 import {getAllAbsentees} from "@/firebase/firebase";
 import { getAge } from "../../common/getAge.js";
-import { getTimeStamp, getLessonDate } from "../../common/date.js";
+import { getTimeStamp, getLessonDate, sortDates } from "../../common/date.js";
 
 export default Vue.extend({
   name: "Absentees",
@@ -135,16 +135,8 @@ export default Vue.extend({
       this.rows = filteredItems.length;
       this.currentPage = 1;
     },
-    sortDates(aRow, bRow, key) {
-      const a = aRow[key];
-      const b = bRow[key];
-
-      if (!a || !b) {
-        return -1;
-      }
-      const dateA = a.split('-').join('');
-      const dateB = b.split('-').join('');
-      return dateA < dateB ? -1 : dateA > dateB ? 1 : 0;
+    sort(aRow, bRow, key) {
+      sortDates(aRow, bRow, key);
     }
   }
 })
